requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
jinja2==3.1.4
python-multipart==0.0.12
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
anthropic>=0.34.2
aiofiles==24.1.0
httpx==0.27.2
python-dotenv==1.0.1
pillow==10.4.0
openpyxl==3.1.5
